Added BulletModel.

+++ Asteroids/src/models/AsteroidModel.java
@@ -2,7 +2,7 @@ package models;
 
 public class AsteroidModel extends MovingObjectModel implements Updatable {
 	
-	public AsteroidModel(int x, int y, int angle, double vx, double vy, double vAngle){
+	public AsteroidModel(int x, int y, double angle, double vx, double vy, double vAngle){
 		super(x, y, angle, vx, vy, vAngle);
 		acc[0] = 0;
 		acc[1] = 0;
