+++ Asteroids/src/models/RogueSpaceshipModel.java
@@ -2,7 +2,7 @@ package models;
 
 public class RogueSpaceshipModel extends MovingObjectModel implements Updatable {
 	
-	public RogueSpaceshipModel(int x, int y, int angle, double vx, double vy, double vAngle){
+	public RogueSpaceshipModel(int x, int y, double angle, double vx, double vy, double vAngle){
 		super(x,y,angle,vx,vy,vAngle);
 	}
 }
